requirements.txt
aiofiles==23.2.1
annotated-types==0.7.0
anthropic==0.45.2
anyio==4.8.0
asgiref==3.8.1
asttokens==2.4.1
asyncpg==0.30.0
cachetools==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
cohere==5.13.12
colorama==0.4.6
Deprecated==1.2.18
devtools==0.12.2
distro==1.9.0
eval_type_backport==0.2.2
executing==2.2.0
fastapi==0.115.8
fastavro==1.10.0
ffmpy==0.5.0
filelock==3.17.0
fsspec==2025.2.0
google-auth==2.38.0
googleapis-common-protos==1.67.0
gradio==5.16.1
gradio_client==1.7.0
griffe==1.5.7
groq==0.18.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.28.1
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
jiter==0.8.2
jsonpath-python==1.0.6
logfire==3.5.3
logfire-api==3.5.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mistralai==1.5.0
mypy-extensions==1.0.0
numpy==2.2.3
openai==1.63.2
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-instrumentation==0.51b0
opentelemetry-instrumentation-asgi==0.51b0
opentelemetry-instrumentation-asyncpg==0.51b0
opentelemetry-instrumentation-dbapi==0.51b0
opentelemetry-instrumentation-fastapi==0.51b0
opentelemetry-instrumentation-sqlite3==0.51b0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opentelemetry-util-http==0.51b0
orjson==3.10.15
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic-ai==0.0.24
pydantic-ai-slim==0.0.24
pydantic-graph==0.0.24
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
redis==5.2.1
requests==2.32.3
rich==13.9.4
rsa==4.9
ruff==0.9.6
safehttpx==0.1.6
semantic-version==2.10.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.45.3
tokenizers==0.21.0
tomlkit==0.13.2
tqdm==4.67.1
typer==0.15.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
websockets==14.2
wrapt==1.17.2
zipp==3.21.0
